refactor: start on type checking + matching api
