+++ apps/frontend/src/composables/servers/modules/network.ts
@@ -1,4 +1,4 @@
-import { pyroFetch } from "../pyro-fetch";
+import { pyroFetch } from "../pyro-fetch.ts";
 import { ServerModule } from "./base";
 import type { Allocation } from "@modrinth/utils";
