+++ packages/utils/servers/errors/index.ts
@@ -1,3 +1,3 @@
-export * from "./pyro-server-error";
+export * from "./pyro-multiple-errors";
 export * from './pyro-fetch-error';
-export * from "./server-error"
+export * from "./pyro-server-error"
